+++ geodata/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class GeodataConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'geodata'
+    verbose_name = 'Геоданные'