+++ restaurant/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class RestaurantConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    verbose_name = 'Рестораны'
     name = 'restaurant'
+    verbose_name = 'Рестораны'