+++ restaurant/v1/views.py
@@ -1,5 +1,3 @@
-from typing import Any
-
 from django.db.models.query import QuerySet
 from django.shortcuts import get_object_or_404
 
@@ -17,12 +15,12 @@ import restaurant.v1.serializers as serializers
 
 @extend_schema(tags=['restaurants'])
 @extend_schema_view(
-    list=extend_schema(description='List all restaurants'),
-    retrieve=extend_schema(description='Get restaurant by ID'),
-    create=extend_schema(description='Create restaurant'),
-    update=extend_schema(description='Update restaurant by ID'),
-    partial_update=extend_schema(description='Partially update restaurant by ID'),
-    destroy=extend_schema(description='Delete restaurant by ID'),
+    list=extend_schema(description='List all restaurants.'),
+    retrieve=extend_schema(description='Get restaurant by ID.'),
+    create=extend_schema(description='Create restaurant.'),
+    update=extend_schema(description='Update restaurant by ID.'),
+    partial_update=extend_schema(description='Partially update restaurant by ID.'),
+    destroy=extend_schema(description='Delete restaurant by ID.'),
     scrape_menu=extend_schema(
         description='Scrape restaurant menu',
         request=OpenApiRequest(OpenApiTypes.NONE),
@@ -45,35 +43,35 @@ class RestaurantViewSet(viewsets.ModelViewSet):
         return self.queryset
 
     @action(detail=True, methods=['post'], url_path='scrape_menu')
-    def scrape_menu(self, request: Request, *args: Any, **kwargs: Any) -> Response:
-        restaurant = get_object_or_404(self.model, pk=kwargs['pk'])
+    def scrape_menu(self, request: Request, pk: int) -> Response:
+        restaurant = get_object_or_404(self.model, pk=pk)
         tasks.scrape_menu_task.delay(restaurant.id)
         return Response(
-            {'status': 'success', 'message': 'Menu scraping task added to queue'},
+            data={'status': 'success', 'message': 'Menu scraping task added to queue'},
             status=status.HTTP_202_ACCEPTED,
         )
 
     @action(detail=True, methods=['post'], url_path='scrape_restaurant')
-    def scrape_restaurant(self, request: Request, *args: Any, **kwargs: Any) -> Response:
-        restaurant = get_object_or_404(self.model, pk=kwargs['pk'])
+    def scrape_restaurant(self, request: Request, pk: int) -> Response:
+        restaurant = get_object_or_404(self.model, pk=pk)
         tasks.scrape_restaurant_task.delay(restaurant.id)
         return Response(
-            {'status': 'success', 'message': 'Restaurant scraping task added to queue'},
+            data={'status': 'success', 'message': 'Restaurant scraping task added to queue'},
             status=status.HTTP_202_ACCEPTED,
         )
 
 
 @extend_schema(tags=['dishes'])
 @extend_schema_view(
-    list=extend_schema(description='List all dishes'),
-    retrieve=extend_schema(description='Get dish by ID'),
-    create=extend_schema(description='Create dish'),
-    update=extend_schema(description='Update dish by ID'),
-    partial_update=extend_schema(description='Partially update dish by ID'),
-    destroy=extend_schema(description='Delete dish by ID'),
+    list=extend_schema(description='List all dishes.'),
+    retrieve=extend_schema(description='Get dish by ID.'),
+    create=extend_schema(description='Create dish.'),
+    update=extend_schema(description='Update dish by ID.'),
+    partial_update=extend_schema(description='Partially update dish by ID.'),
+    destroy=extend_schema(description='Delete dish by ID.'),
 )
 class DishViewSet(viewsets.ModelViewSet):
     serializer_class = serializers.DishSerializer
     model = serializer_class.Meta.model
-    queryset = model.objects.select_related('restaurant')
+    queryset = model.objects.select_related('restaurant').order_by('pk')
     filterset_class = DishFilterSet